requirements.txt
Flask==2.0.1
werkzeug==2.0.1
pyTelegramBotAPI==4.11.0
requests==2.28.2
pandas==1.5.3
numpy==1.23.5
pandas==1.5.3
python-dotenv
